Organize the read of the databases
The reading of the databases was re-organized, because it was
a big mess. Now, there is a reading of billing databases (now
in a properly named folder), and buy orders databases. The reading
steps were broken down into 3: retrieve of files to read, population
of the dictionary with every table of the databases and the reading
of every specific database (6 new functions in total).